Remove 'isolated' module from finance. Move the dummy contract to test-utils and create a task for generating a jar containing it.
