--- finance/isolated/build.gradle
@@ -1,6 +1,0 @@
-apply plugin: 'kotlin'
-apply plugin: CanonicalizerPlugin
-
-dependencies {
-    compile project(':core')
-}
