+++ test-utils/src/main/kotlin/net/corda/testing/contracts/AnotherDummyContract.kt
@@ -1,4 +1,4 @@
-package net.corda.contracts
+package net.corda.testing.contracts
 
 import net.corda.core.contracts.*
 import net.corda.core.crypto.SecureHash
@@ -11,7 +11,7 @@ import net.corda.core.transactions.TransactionBuilder
 
 val ANOTHER_DUMMY_PROGRAM_ID = AnotherDummyContract()
 
-class AnotherDummyContract : Contract, net.corda.core.node.DummyContractBackdoor {
+class AnotherDummyContract : Contract, DummyContractBackdoor {
     data class State(val magicNumber: Int = 0) : ContractState {
         override val contract = ANOTHER_DUMMY_PROGRAM_ID
         override val participants: List<AbstractParty>
