+++ test-utils/src/main/kotlin/net/corda/testing/contracts/DummyContractBackdoor.kt
@@ -1,4 +1,4 @@
-package net.corda.core.node
+package net.corda.testing.contracts
 
 import net.corda.core.contracts.ContractState
 import net.corda.core.contracts.PartyAndReference
@@ -7,6 +7,5 @@ import net.corda.core.transactions.TransactionBuilder
 
 interface DummyContractBackdoor {
     fun generateInitial(owner: PartyAndReference, magicNumber: Int, notary: Party): TransactionBuilder
-
     fun inspectState(state: ContractState): Int
 }
